+++ sop-common/sop-gateway-common/src/main/java/com/gitee/sop/gatewaycommon/gateway/filter/IndexFilter.java
@@ -20,6 +20,7 @@ import org.springframework.core.annotation.Order;
 import org.springframework.core.io.buffer.DataBuffer;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
+import org.springframework.http.codec.ServerCodecConfigurer;
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.http.server.reactive.ServerHttpRequestDecorator;
 import org.springframework.web.reactive.function.BodyInserter;
@@ -44,6 +45,9 @@ import java.util.Objects;
 @Slf4j
 @Order(Ordered.HIGHEST_PRECEDENCE)
 public class IndexFilter implements WebFilter {
+
+    @Autowired
+    private ServerCodecConfigurer codecConfigurer;
 
     /** 路径白名单 */
     private static final List<String> PATH_WHITE_LIST = Arrays.asList(
@@ -78,7 +82,7 @@ public class IndexFilter implements WebFilter {
         }
         if (Objects.equals(path, indexPath) || "".equals(path)) {
             if (request.getMethod() == HttpMethod.POST) {
-                ServerRequest serverRequest = ServerWebExchangeUtil.createReadBodyRequest(exchange);
+                ServerRequest serverRequest = ServerWebExchangeUtil.createReadBodyRequest(exchange, codecConfigurer);
                 // 读取请求体中的内容
                 Mono<?> modifiedBody = serverRequest.bodyToMono(byte[].class)
                         .switchIfEmpty(Mono.just("".getBytes()))