+++ changelog.md
@@ -1,5 +1,10 @@
 # changelog
 
+## 4.3.0
+
+- 升级`spring-boot/spring-cloud/spring-cloud-alibaba`版本
+- 修复`DataBufferLimitException: Exceeded limit on max bytes to buffer : 262144`问题
+
 ## 4.2.7
 
 - 修复两个微服务相同path问题